+++ CHANGELOG.md
@@ -25,6 +25,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Algo creation events aren't included in newsfeed anymore.
 
+### Changed
+
+- Remove task category from the compute engine.
+
 ## [0.32.0] 2022-10-03
 
 ### Changed